split files, memory alignment and allocation. remove redundant headers
